+++ atom/browser/api/atom_api_browser_window.h
@@ -5,43 +5,21 @@
 #ifndef ATOM_BROWSER_API_ATOM_API_BROWSER_WINDOW_H_
 #define ATOM_BROWSER_API_ATOM_API_BROWSER_WINDOW_H_
 
-#include <map>
-#include <memory>
 #include <string>
 #include <vector>
 
+#include "atom/browser/api/atom_api_top_level_window.h"
 #include "atom/browser/api/atom_api_web_contents.h"
-#include "atom/browser/native_window.h"
-#include "atom/browser/native_window_observer.h"
-#include "atom/common/api/atom_api_native_image.h"
-#include "atom/common/key_weak_map.h"
 #include "base/cancelable_callback.h"
-#include "base/memory/weak_ptr.h"
-#include "content/public/browser/render_widget_host.h"
-#include "native_mate/persistent_dictionary.h"
-
-class GURL;
-
-namespace gfx {
-class Rect;
-}
-
-namespace mate {
-class Arguments;
-class Dictionary;
-}
 
 namespace atom {
 
-class NativeWindow;
-
 namespace api {
 
-class BrowserWindow : public mate::TrackableObject<BrowserWindow>,
+class BrowserWindow : public TopLevelWindow,
                       public content::RenderWidgetHost::InputEventObserver,
                       public content::WebContentsObserver,
-                      public ExtendedWebContentsObserver,
-                      public NativeWindowObserver {
+                      public ExtendedWebContentsObserver {
  public:
   static mate::WrappableBase* New(mate::Arguments* args);
 
@@ -52,9 +30,9 @@ class BrowserWindow : public mate::TrackableObject<BrowserWindow>,
   static v8::Local<v8::Value> From(v8::Isolate* isolate,
                                    NativeWindow* native_window);
 
-  NativeWindow* window() const { return window_.get(); }
-
-  int32_t ID() const;
+  base::WeakPtr<BrowserWindow> GetWeakPtr() {
+    return weak_factory_.GetWeakPtr();
+  }
 
  protected:
   BrowserWindow(v8::Isolate* isolate,
@@ -80,184 +58,28 @@ class BrowserWindow : public mate::TrackableObject<BrowserWindow>,
   void OnRendererResponsive() override;
 
   // NativeWindowObserver:
-  void WillCloseWindow(bool* prevent_default) override;
   void RequestPreferredWidth(int* width) override;
   void OnCloseButtonClicked(bool* prevent_default) override;
+
+  // TopLevelWindow:
   void OnWindowClosed() override;
-  void OnWindowEndSession() override;
   void OnWindowBlur() override;
   void OnWindowFocus() override;
-  void OnWindowShow() override;
-  void OnWindowHide() override;
-  void OnWindowMaximize() override;
-  void OnWindowUnmaximize() override;
-  void OnWindowMinimize() override;
-  void OnWindowRestore() override;
   void OnWindowResize() override;
-  void OnWindowMove() override;
-  void OnWindowMoved() override;
-  void OnWindowScrollTouchBegin() override;
-  void OnWindowScrollTouchEnd() override;
-  void OnWindowSwipe(const std::string& direction) override;
-  void OnWindowSheetBegin() override;
-  void OnWindowSheetEnd() override;
-  void OnWindowEnterFullScreen() override;
-  void OnWindowLeaveFullScreen() override;
-  void OnWindowEnterHtmlFullScreen() override;
-  void OnWindowLeaveHtmlFullScreen() override;
-  void OnExecuteWindowsCommand(const std::string& command_name) override;
-  void OnTouchBarItemResult(const std::string& item_id,
-                            const base::DictionaryValue& details) override;
-  void OnNewWindowForTab() override;
+  void Focus() override;
+  void Blur() override;
+  void SetBackgroundColor(const std::string& color_name) override;
+  void SetBrowserView(v8::Local<v8::Value> value) override;
+  void SetVibrancy(mate::Arguments* args) override;
 
-  #if defined(OS_WIN)
-  void OnWindowMessage(UINT message, WPARAM w_param, LPARAM l_param) override;
-  #endif
-
-  base::WeakPtr<BrowserWindow> GetWeakPtr() {
-    return weak_factory_.GetWeakPtr();
-  }
-
- private:
-  void Init(v8::Isolate* isolate,
-            v8::Local<v8::Object> wrapper,
-            const mate::Dictionary& options,
-            mate::Handle<class WebContents> web_contents);
-
-  // APIs for NativeWindow.
-  void Close();
-  void Focus();
-  void Blur();
-  bool IsFocused();
-  void Show();
-  void ShowInactive();
-  void Hide();
-  bool IsVisible();
-  bool IsEnabled();
-  void SetEnabled(bool enable);
-  void Maximize();
-  void Unmaximize();
-  bool IsMaximized();
-  void Minimize();
-  void Restore();
-  bool IsMinimized();
-  void SetFullScreen(bool fullscreen);
-  bool IsFullscreen();
-  void SetBounds(const gfx::Rect& bounds, mate::Arguments* args);
-  gfx::Rect GetBounds();
-  void SetSize(int width, int height, mate::Arguments* args);
-  std::vector<int> GetSize();
-  void SetContentSize(int width, int height, mate::Arguments* args);
-  std::vector<int> GetContentSize();
-  void SetContentBounds(const gfx::Rect& bounds, mate::Arguments* args);
-  gfx::Rect GetContentBounds();
-  void SetMinimumSize(int width, int height);
-  std::vector<int> GetMinimumSize();
-  void SetMaximumSize(int width, int height);
-  std::vector<int> GetMaximumSize();
-  void SetSheetOffset(double offsetY, mate::Arguments* args);
-  void SetResizable(bool resizable);
-  bool IsResizable();
-  void SetMovable(bool movable);
-  #if defined(OS_WIN) || defined(OS_MACOSX)
-  void MoveTop();
-  #endif
-  bool IsMovable();
-  void SetMinimizable(bool minimizable);
-  bool IsMinimizable();
-  void SetMaximizable(bool maximizable);
-  bool IsMaximizable();
-  void SetFullScreenable(bool fullscreenable);
-  bool IsFullScreenable();
-  void SetClosable(bool closable);
-  bool IsClosable();
-  void SetAlwaysOnTop(bool top, mate::Arguments* args);
-  bool IsAlwaysOnTop();
-  void Center();
-  void SetPosition(int x, int y, mate::Arguments* args);
-  std::vector<int> GetPosition();
-  void SetTitle(const std::string& title);
-  std::string GetTitle();
-  void FlashFrame(bool flash);
-  void SetSkipTaskbar(bool skip);
-  void SetSimpleFullScreen(bool simple_fullscreen);
-  bool IsSimpleFullScreen();
-  void SetKiosk(bool kiosk);
-  bool IsKiosk();
-  void SetBackgroundColor(const std::string& color_name);
-  void SetHasShadow(bool has_shadow);
-  bool HasShadow();
-  void SetOpacity(const double opacity);
-  double GetOpacity();
+  // BrowserWindow APIs.
   void FocusOnWebView();
   void BlurWebView();
   bool IsWebViewFocused();
-  void SetRepresentedFilename(const std::string& filename);
-  std::string GetRepresentedFilename();
-  void SetDocumentEdited(bool edited);
-  bool IsDocumentEdited();
-  void SetIgnoreMouseEvents(bool ignore, mate::Arguments* args);
-  void SetContentProtection(bool enable);
-  void SetFocusable(bool focusable);
-  void SetProgressBar(double progress, mate::Arguments* args);
-  void SetOverlayIcon(const gfx::Image& overlay,
-                      const std::string& description);
-  bool SetThumbarButtons(mate::Arguments* args);
-  void SetMenu(v8::Isolate* isolate, v8::Local<v8::Value> menu);
-  void SetAutoHideMenuBar(bool auto_hide);
-  bool IsMenuBarAutoHide();
-  void SetMenuBarVisibility(bool visible);
-  bool IsMenuBarVisible();
-  void SetAspectRatio(double aspect_ratio, mate::Arguments* args);
-  void PreviewFile(const std::string& path, mate::Arguments* args);
-  void CloseFilePreview();
-  void SetParentWindow(v8::Local<v8::Value> value, mate::Arguments* args);
-  v8::Local<v8::Value> GetParentWindow() const;
-  std::vector<v8::Local<v8::Object>> GetChildWindows() const;
-  v8::Local<v8::Value> GetBrowserView() const;
-  void SetBrowserView(v8::Local<v8::Value> value);
-  void ResetBrowserView();
-  bool IsModal() const;
-  v8::Local<v8::Value> GetNativeWindowHandle();
+  v8::Local<v8::Value> GetWebContents(v8::Isolate* isolate);
 
-#if defined(OS_WIN)
-  typedef base::Callback<void(v8::Local<v8::Value>,
-                              v8::Local<v8::Value>)> MessageCallback;
-
-  bool HookWindowMessage(UINT message, const MessageCallback& callback);
-  bool IsWindowMessageHooked(UINT message);
-  void UnhookWindowMessage(UINT message);
-  void UnhookAllWindowMessages();
-  bool SetThumbnailClip(const gfx::Rect& region);
-  bool SetThumbnailToolTip(const std::string& tooltip);
-  void SetAppDetails(const mate::Dictionary& options);
-#endif
-
-#if defined(TOOLKIT_VIEWS)
-  void SetIcon(mate::Handle<NativeImage> icon);
-#endif
-
-  void SetVisibleOnAllWorkspaces(bool visible);
-  bool IsVisibleOnAllWorkspaces();
-
-  void SetAutoHideCursor(bool auto_hide);
-
-  void SelectPreviousTab();
-  void SelectNextTab();
-  void MergeAllWindows();
-  void MoveTabToNewWindow();
-  void ToggleTabBar();
-  void AddTabbedWindow(NativeWindow* window, mate::Arguments* args);
-
-  void SetVibrancy(mate::Arguments* args);
-  void SetTouchBar(const std::vector<mate::PersistentDictionary>& items);
-  void RefreshTouchBarItem(const std::string& item_id);
-  void SetEscapeTouchBarItem(const mate::PersistentDictionary& item);
-
-  v8::Local<v8::Value> WebContents(v8::Isolate* isolate);
-
-  // Remove this window from parent window's |child_windows_|.
-  void RemoveFromParentChildWindows();
+ private:
+  // Helpers.
 
   // Called when the window needs to update its draggable region.
   void UpdateDraggableRegions(
@@ -274,28 +96,19 @@ class BrowserWindow : public mate::TrackableObject<BrowserWindow>,
   // Dispatch unresponsive event to observers.
   void NotifyWindowUnresponsive();
 
-#if defined(OS_WIN)
-  typedef std::map<UINT, MessageCallback> MessageCallbackMap;
-  MessageCallbackMap messages_callback_map_;
-#endif
-
-#if defined(OS_MACOSX)
-  std::vector<DraggableRegion> draggable_regions_;
-#endif
+  // Cleanup our WebContents observers.
+  void Cleanup();
 
   // Closure that would be called when window is unresponsive when closing,
   // it should be cancelled when we can prove that the window is responsive.
   base::CancelableClosure window_unresponsive_closure_;
 
-  v8::Global<v8::Value> browser_view_;
+#if defined(OS_MACOSX)
+  std::vector<DraggableRegion> draggable_regions_;
+#endif
+
   v8::Global<v8::Value> web_contents_;
-  v8::Global<v8::Value> menu_;
-  v8::Global<v8::Value> parent_window_;
-  KeyWeakMap<int> child_windows_;
-
   api::WebContents* api_web_contents_;
-
-  std::unique_ptr<NativeWindow> window_;
 
   base::WeakPtrFactory<BrowserWindow> weak_factory_;
 
@@ -305,7 +118,6 @@ class BrowserWindow : public mate::TrackableObject<BrowserWindow>,
 }  // namespace api
 
 }  // namespace atom
-
 
 namespace mate {
