+++ filenames.gypi
@@ -152,6 +152,8 @@
       'atom/browser/api/atom_api_system_preferences.h',
       'atom/browser/api/atom_api_system_preferences_mac.mm',
       'atom/browser/api/atom_api_system_preferences_win.cc',
+      'atom/browser/api/atom_api_top_level_window.cc',
+      'atom/browser/api/atom_api_top_level_window.h',
       'atom/browser/api/atom_api_tray.cc',
       'atom/browser/api/atom_api_tray.h',
       'atom/browser/api/atom_api_url_request.cc',
